[STDIO] Split printf.c into two files
